Bugfix for incorrect logic_is_dl() results
When called for the first time, logic_is_dl() was looking for the
set-info command in commands[0]. However, the parse loop in main()
only reads the benchmark up to its first check-sat command; then this
part is printed and the commands vector is cleared. Thus, on
benchmarks where the first check-sat command is encountered before the
first call to logic_is_dl(), the latter would no longer be able to
determine the correct logic, but would (incorrectly, if the logic was
in fact a difference logic) return false.