Major refactoring: name ids no longer resolve shadowing
Any shadowing of names present in the input benchmark will cause
corresponding shadowing in the scrambled output.

This allows us to simplify the implementation by removing the symbol
table and associated scope management. Instead, the scrambler merely
builds a bijection between input names and name ids (from which
uniform output names of the form x1, x2, ... are derived).

This leads to an order of magnitude speedup on certain incremental
benchmarks in SMT-LIB, reducing scrambling time from ~ 30 minutes
(which is infeasible in the context of SMT-COMP) to ~ 3 minutes.