Add scoring for competition-wide recognitions.
Previously the score function computed the scores per solver in a
division. Since we have to determine the VBS on a per benchmark basis
for the Largest Contribution Ranking, the score function now computes
the scores per benchmark.

This commit further refactors large parts of the script to speed-up the
computations.