Model Validation scoring
As per rules, all invalid models are reported as errors.
Note, that missing model or missing sat result is reported as
unknown.

Note further that this is for this year only.  For the next year only
unsound models and unsat results should be reported as error, as we
plan to change the rules.  Syntactic problems in the model are then
reported as unknown.